Fix sending SMS
We used to cast the phone numbers to int, that may give 0 when the
string was not numeric. Rewrote a regex to get all numeric characters
from the phone number string.

Because we send the message and phone numbers in
GET parameters to clickatell, I decreased the chunk limit to 5 at a
time.